+++ packages/react/src/blocks/forms/TextArea.tsx
@@ -8,6 +8,7 @@ export interface FormTextAreaProps {
   value?: string;
   defaultValue?: string;
   placeholder?: string;
+  required?: boolean;
 }
 
 class TextAreaComponent extends React.Component<FormTextAreaProps> {
@@ -18,6 +19,7 @@ class TextAreaComponent extends React.Component<FormTextAreaProps> {
         name={this.props.name}
         value={this.props.value}
         defaultValue={this.props.defaultValue}
+        required={this.props.required}
         {...this.props.attributes}
       />
     );