+++ packages/sdks/src/blocks/embed/component-info.ts
@@ -4,6 +4,12 @@ import type { ComponentInfo } from '../../types/components.js';
 declare const _iframelyApiKey: string;
 
 export const componentInfo: ComponentInfo = {
+  shouldReceiveBuilderProps: {
+    builderBlock: false,
+    builderContext: false,
+    builderComponents: false,
+    builderLinkComponent: false,
+  },
   name: 'Embed',
   static: true,