+++ packages/sdks/src/blocks/form/form/component-info.ts
@@ -279,4 +279,11 @@ export const componentInfo: ComponentInfo = {
       },
     },
   ],
+
+  shouldReceiveBuilderProps: {
+    builderBlock: true,
+    builderContext: true,
+    builderComponents: true,
+    builderLinkComponent: true,
+  },
 };