+++ packages/sdks/src/blocks/raw-text/component-info.ts
@@ -1,6 +1,12 @@
 import type { ComponentInfo } from '../../types/components.js';
 
 export const componentInfo: ComponentInfo = {
+  shouldReceiveBuilderProps: {
+    builderBlock: false,
+    builderContext: false,
+    builderComponents: false,
+    builderLinkComponent: false,
+  },
   name: 'Builder:RawText',
   hideFromInsertMenu: true,