+++ packages/sdks/src/functions/evaluate/node-runtime/node-runtime.ts
@@ -1,4 +1,5 @@
 import type { Context, IsolateOptions } from 'isolated-vm';
+import { SDK_NAME } from '../../../constants/sdk-name.js';
 import { MSG_PREFIX, logger } from '../../../helpers/logger.js';
 import { fastClone } from '../../fast-clone.js';
 import { set } from '../../set.js';
@@ -90,6 +91,11 @@ export const setIvm = (ivm: IsolatedVMImport, options: IsolateOptions = {}) => {
   setIsolateContext(options);
 };
 
+// only mention the script for SDKs that have it.
+const SHOULD_MENTION_INITIALIZE_SCRIPT =
+  SDK_NAME === '@builder.io/sdk-react-nextjs' ||
+  SDK_NAME === '@builder.io/sdk-react';
+
 const getIvm = (): IsolatedVMImport => {
   try {
     if (IVM_INSTANCE) return IVM_INSTANCE;
@@ -100,16 +106,16 @@ const getIvm = (): IsolatedVMImport => {
     logger.error('isolated-vm import error.', error);
   }
 
-  throw new Error(
-    `${MSG_PREFIX}could not import \`isolated-vm\` module for safe script execution on Node server.
+  const ERROR_MESSAGE = `${MSG_PREFIX}could not import \`isolated-vm\` module for safe script execution on a Node server.
     
-    In certain Node environments, the SDK requires additional initialization steps. This can be achieved by 
-    importing and calling \`initializeNodeRuntime()\` from "@builder.io/sdk-react/node/init". This must be done in
-    a server-only execution path within your application.
+    SOLUTION: In a server-only execution path within your application, do one of the following:
+  
+    ${SHOULD_MENTION_INITIALIZE_SCRIPT ? '- import and call `initializeNodeRuntime()` from "${SDK_NAME}/node/init".' : ''}
+    - add the following import: \`await import('isolated-vm')\`.
 
-    Please see the documentation for more information: https://builder.io/c/docs/integration-tips#enabling-data-bindings-in-node-environments
-    `
-  );
+    For more information, visit https://builder.io/c/docs/integration-tips#enabling-data-bindings-in-node-environments`;
+
+  throw new Error(ERROR_MESSAGE);
 };
 
 function setIsolateContext(options: IsolateOptions = { memoryLimit: 128 }) {